refactor mocks path so gradle can build
