+++ Flint/src/test/java/mocks/NotConfig.java
@@ -1,4 +1,4 @@
-package flint;
+package mocks;
 
 public class NotConfig {
 }
