+++ Flint/src/test/java/mocks/TestConfig.java
@@ -1,4 +1,6 @@
-package flint;
+package mocks;
+
+import flint.FlintConfiguration;
 
 public class TestConfig extends FlintConfiguration {
 }
